added lombok , refactor controllers and use actuator as health check
