+++ build.gradle
@@ -38,6 +38,9 @@ dependencies {
 	implementation group: 'com.google.code.gson', name: 'gson', version: '2.7'
 
 	testCompile group: 'junit', name: 'junit'
+
+	compileOnly 'org.projectlombok:lombok:1.18.4'
+	annotationProcessor 'org.projectlombok:lombok:1.18.4'
 }
 
 test {
