+++ src/main/java/utn/tacs/common/client/superHeroAPI/clientApi/model/Character.java
@@ -1,28 +1,20 @@
 package utn.tacs.common.client.superHeroAPI.clientApi.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.io.Serializable;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Character {
+@Getter @Setter
+@NoArgsConstructor
+public class Character implements Serializable {
 
     private String id;
     private String name;
     private Powerstats powerstats;
     private Image image;
 
-    public String getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Powerstats getPowerstats() {
-        return powerstats;
-    }
-
-    public Image getImage() {
-        return image;
-    }
 }
