+++ src/main/java/utn/tacs/common/client/superHeroAPI/clientApi/model/Image.java
@@ -1,13 +1,16 @@
 package utn.tacs.common.client.superHeroAPI.clientApi.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
+@Getter
+@Setter
+@NoArgsConstructor
 public class Image {
 
     private String url;
 
-    public String getUrl() {
-        return url;
-    }
 }
