+++ src/main/java/utn/tacs/controllers/MatchController.java
@@ -1,20 +1,26 @@
-package utn.tacs.controllers.matches;
+package utn.tacs.controllers;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import org.springframework.lang.NonNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import utn.tacs.dto.battle.BattleModelResponse;
+import utn.tacs.dto.battle.MatchBattleRequest;
 import utn.tacs.dto.card.CardModelResponse;
-import utn.tacs.dto.match.MatchDrawRequest;
-import utn.tacs.dto.match.MatchFindRequest;
-import utn.tacs.dto.match.MatchModelResponse;
-import utn.tacs.dto.match.MatchPagingRequest;
+import utn.tacs.dto.deck.response.MatchModel;
+import utn.tacs.dto.deck.response.Request;
+import utn.tacs.dto.match.*;
+import utn.tacs.services.MatchBattleHandler;
 import utn.tacs.services.MatchCardDrawer;
+import utn.tacs.services.MatchesCreator;
 import utn.tacs.services.MatchesFinder;
 
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -23,14 +29,18 @@ import java.util.List;
 @Api(tags = "Matches")
 @CrossOrigin("*")
 @RestController
-public class MatchesGetController {
+public class MatchController {
 
-    MatchesFinder finder;
-    MatchCardDrawer cardDraw;
+    private MatchesFinder finder;
+    private MatchCardDrawer cardDraw;
+    private MatchesCreator creator;
+    private MatchBattleHandler battle;
 
-    public MatchesGetController(MatchesFinder finder, MatchCardDrawer cardDraw) {
+    public MatchController(MatchesFinder finder, MatchCardDrawer cardDraw, MatchesCreator creator, MatchBattleHandler battle) {
         this.finder = finder;
         this.cardDraw = cardDraw;
+        this.creator = creator;
+        this.battle = battle;
     }
 
     @GetMapping
@@ -64,6 +74,33 @@ public class MatchesGetController {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         String playerId = auth.getName();
         return cardDraw.draw(new MatchDrawRequest(id, playerId));
+    }
+
+    @PostMapping
+    @ApiOperation(value = "Crear partida, se divide y mezcla el mazo elegido")
+    @ApiResponses({
+            @ApiResponse(code = 200, response = MatchModelResponse.class, message = "match")
+    })
+    public MatchModelResponse postMatch(@Validated @NonNull @RequestBody MatchModel matchRequest) throws Exception {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String hostId = auth.getName();
+
+        List<String> players = new ArrayList<>();
+        players.add(hostId);
+        players.add(matchRequest.getOpponentId());
+
+        return creator.create(new MatchCreateRequest(players, matchRequest.getDeckId()));
+    }
+
+
+    @PostMapping("/{id}/battles")
+    @ApiOperation(value = "Indicar en tu turno que atributo se usara")
+    @ApiResponses({
+            @ApiResponse(code = 200, response = BattleModelResponse.class, message = "Resultado del combate")
+    })
+    public BattleModelResponse modifyMatch(@RequestBody Request request, @PathVariable("id") String id ) throws Exception {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return battle.begin(new MatchBattleRequest(id, auth.getName(), request.getAttribute()));
     }
 
 }
