+++ src/main/java/utn/tacs/controllers/UsersGetController.java
@@ -1,4 +1,4 @@
-package utn.tacs.controllers.users;
+package utn.tacs.controllers;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
