+++ src/main/java/utn/tacs/domain/Battle.java
@@ -1,5 +1,8 @@
 package utn.tacs.domain;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.common.client.superHeroAPI.clientApi.SuperHeroApi;
 import utn.tacs.common.client.superHeroAPI.clientApi.model.Powerstats;
 import utn.tacs.dto.deck.response.Attribute;
@@ -8,7 +11,11 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Queue;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class Battle {
+
     private Attribute attribute;
     private Map<String, Card> players;
     private String winner;
@@ -36,15 +43,4 @@ public class Battle {
         }
     }
 
-    public String getWinner(){
-        return winner;
-    }
-
-    public Attribute getAttribute() {
-        return attribute;
-    }
-
-    public Map<String, Card> getPlayers() {
-        return players;
-    }
 }
