+++ src/main/java/utn/tacs/domain/CardId.java
@@ -1,22 +1,19 @@
 package utn.tacs.domain;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 import java.util.Objects;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class CardId {
 
     private String id;
 
-    public CardId(){}
-
     public CardId(String id) {
-        this.id = id;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
         this.id = id;
     }
