+++ src/main/java/utn/tacs/domain/PowerStats.java
@@ -1,11 +1,18 @@
 package utn.tacs.domain;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.dto.deck.response.Attribute;
 
 import java.util.HashMap;
 import java.util.Map;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class PowerStats {
+
     private final Map<Attribute, Integer> attributes = new HashMap<>();
 
     public PowerStats addAttribute(Attribute attribute, Integer value){
@@ -15,9 +22,5 @@ public class PowerStats {
 
     public Integer getAttribute(Attribute attribute){
         return attributes.get(attribute);
-    }
-
-    public Map<Attribute, Integer> getAttributes() {
-        return attributes;
     }
 }
