+++ src/main/java/utn/tacs/dto/battle/BattleModelResponse.java
@@ -1,42 +1,25 @@
 package utn.tacs.dto.battle;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.domain.Card;
 import utn.tacs.dto.deck.response.Attribute;
 
 import java.util.Map;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class BattleModelResponse {
-    Attribute attribute;
-    Map<String, Card> players;
-    String winner;
+
+    private Attribute attribute;
+    private Map<String, Card> players;
+    private String winner;
 
     public BattleModelResponse(Attribute attribute, Map<String, Card> players, String winner) {
         this.attribute = attribute;
         this.players = players;
-        this.winner = winner;
-    }
-
-    public Attribute getAttribute() {
-        return attribute;
-    }
-
-    public void setAttribute(Attribute attribute) {
-        this.attribute = attribute;
-    }
-
-    public Map<String, Card> getPlayers() {
-        return players;
-    }
-
-    public void setPlayers(Map<String, Card> players) {
-        this.players = players;
-    }
-
-    public String getWinner() {
-        return winner;
-    }
-
-    public void setWinner(String winner) {
         this.winner = winner;
     }
 }
