+++ src/main/java/utn/tacs/dto/battle/MatchBattleRequest.java
@@ -1,39 +1,22 @@
 package utn.tacs.dto.battle;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.dto.deck.response.Attribute;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class MatchBattleRequest {
-    String matchId;
-    String playerId;
-    Attribute attribute;
+
+    private String matchId;
+    private String playerId;
+    private Attribute attribute;
 
     public MatchBattleRequest(String matchId, String playerId, Attribute attribute) {
         this.matchId = matchId;
         this.playerId = playerId;
-        this.attribute = attribute;
-    }
-
-    public String getMatchId() {
-        return matchId;
-    }
-
-    public void setMatchId(String matchId) {
-        this.matchId = matchId;
-    }
-
-    public String getPlayerId() {
-        return playerId;
-    }
-
-    public void setPlayerId(String playerId) {
-        this.playerId = playerId;
-    }
-
-    public Attribute getAttribute() {
-        return attribute;
-    }
-
-    public void setAttribute(Attribute attribute) {
         this.attribute = attribute;
     }
 }
