+++ src/main/java/utn/tacs/dto/card/CardFindRequest.java
@@ -1,19 +1,19 @@
 package utn.tacs.dto.card;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.domain.CardId;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class CardFindRequest {
+
     private CardId cardId;
 
     public CardFindRequest(CardId cardId) {
         this.cardId = cardId;
     }
 
-    public CardId getCardId() {
-        return cardId;
-    }
-
-    public void setCardId(CardId cardId) {
-        this.cardId = cardId;
-    }
 }
