+++ src/main/java/utn/tacs/dto/deck/DeckCreateRequest.java
@@ -1,9 +1,15 @@
 package utn.tacs.dto.deck;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.domain.CardId;
 
 import java.util.List;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class DeckCreateRequest {
 
     private List<CardId> cardIds;
@@ -12,13 +18,5 @@ public class DeckCreateRequest {
     public DeckCreateRequest(List<CardId> cardIds, String name) {
         this.cardIds = cardIds;
         this.name = name;
-    }
-
-    public List<CardId> getCards() {
-        return cardIds;
-    }
-
-    public String getName(){
-        return name;
     }
 }
