+++ src/main/java/utn/tacs/dto/deck/DeckUpdateRequest.java
@@ -1,32 +1,23 @@
 package utn.tacs.dto.deck;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 import java.util.List;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class DeckUpdateRequest {
 
-    private final String deckId;
-    private final String newName;
+    private String deckId;
+    private String newName;
     private List<String> cards;
 
     public DeckUpdateRequest(String deckId, String newName, List<String> cards) {
         this.deckId = deckId;
         this.newName = newName;
-        this.cards = cards;
-    }
-
-    public String getDeckId() {
-        return deckId;
-    }
-
-    public String getNewName() {
-        return newName;
-    }
-
-    public List<String> getCards() {
-        return cards;
-    }
-
-    public void setCards(List<String> cards) {
         this.cards = cards;
     }
 }
