+++ src/main/java/utn/tacs/dto/deck/response/Request.java
@@ -1,20 +1,13 @@
 package utn.tacs.dto.deck.response;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
 public class Request {
 
     private Attribute attribute;
-
-    public Request(){}
-
-    public Request(Attribute attribute) {
-        this.attribute = attribute;
-    }
-
-    public Attribute getAttribute() {
-        return attribute;
-    }
-
-    public void setAttribute(Attribute attribute) {
-        this.attribute = attribute;
-    }
 }
