+++ src/main/java/utn/tacs/dto/deck/response/UserModelRespones.java
@@ -1,5 +1,12 @@
 package utn.tacs.dto.deck.response;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
 public class UserModelRespones {
 
     private String id;
