+++ src/main/java/utn/tacs/dto/match/MatchFindRequest.java
@@ -1,18 +1,17 @@
 package utn.tacs.dto.match;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
 public class MatchFindRequest {
 
     private String matchId;
 
     public MatchFindRequest(String matchId) {
-        this.matchId = matchId;
-    }
-
-    public String getMatchId() {
-        return matchId;
-    }
-
-    public void setMatchId(String matchId) {
         this.matchId = matchId;
     }
 }
