+++ src/main/java/utn/tacs/dto/match/MatchModelResponse.java
@@ -1,6 +1,9 @@
 package utn.tacs.dto.match;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import utn.tacs.domain.Battle;
 import utn.tacs.domain.Match;
 import utn.tacs.dto.deck.response.MatchStatusEnum;
@@ -10,6 +13,9 @@ import java.util.Date;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
+@Getter
+@Setter
+@NoArgsConstructor
 public class MatchModelResponse {
 
     private String id;
@@ -35,69 +41,5 @@ public class MatchModelResponse {
         matchModelResponse.setEndDate(match.getEndDate());
         matchModelResponse.setWinnerID(match.getWinnerID());
         return matchModelResponse;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public List<String> getPlayers() {
-        return players;
-    }
-
-    public void setPlayers(List<String> players) {
-        this.players = players;
-    }
-
-    public String getDeck() {
-        return deck;
-    }
-
-    public void setDeck(String deck) {
-        this.deck = deck;
-    }
-
-    public MatchStatusEnum getStatus() {
-        return status;
-    }
-
-    public void setStatus(MatchStatusEnum status) {
-        this.status = status;
-    }
-
-    public Date getCreationDate() {
-        return creationDate;
-    }
-
-    public void setCreationDate(Date creationDate) {
-        this.creationDate = creationDate;
-    }
-
-    public Date getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(Date endDate) {
-        this.endDate = endDate;
-    }
-
-    public String getWinnerID() {
-        return winnerID;
-    }
-
-    public void setWinnerID(String winnerID) {
-        this.winnerID = winnerID;
-    }
-
-    public List<Battle> getBattles() {
-        return battles;
-    }
-
-    public void setBattles(List<Battle> battles) {
-        this.battles = battles;
     }
 }
