+++ src/main/java/utn/tacs/dto/match/Paging.java
@@ -1,5 +1,12 @@
 package utn.tacs.dto.match;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
 public abstract class Paging {
 
     private int page;
@@ -9,30 +16,6 @@ public abstract class Paging {
     public Paging(int page, int pageSize, String sortDirection) {
         this.page = page;
         this.pageSize = pageSize;
-        this.sortDirection = sortDirection;
-    }
-
-    public int getPage() {
-        return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
-    }
-
-    public int getPageSize() {
-        return pageSize;
-    }
-
-    public void setPageSize(int pageSize) {
-        this.pageSize = pageSize;
-    }
-
-    public String getSortDirection() {
-        return sortDirection;
-    }
-
-    public void setSortDirection(String sortDirection) {
         this.sortDirection = sortDirection;
     }
 }
