+++ src/main/java/utn/tacs/services/DecksCreator.java
@@ -16,8 +16,8 @@ public class DecksCreator {
     }
 
     public DeckModelResponse create(DeckCreateRequest deckCreateRequest){
-        Deck deck = new Deck(deckCreateRequest.getCards(), deckCreateRequest.getName());
+        Deck deck = new Deck(deckCreateRequest.getCardIds(), deckCreateRequest.getName());
         repository.save(deck);
-        return new DeckModelResponse(deck.getCards(), deck.getId(), deck.getName());
+        return new DeckModelResponse(deck.getCardIds(), deck.getId(), deck.getName());
     }
 }
