+++ src/main/resources/application.properties
@@ -2,6 +2,6 @@ server.port=8080
 auth0.audience=https://tacs.2021.com
 auth0.domain=dev-jx8fysvq.us.auth0.com
 spring.security.oauth2.resourceserver.jwt.issuer-uri=https://dev-jx8fysvq.us.auth0.com/
-
+springdoc.show.actuator=true
 #ehcache
 spring.cache.jcache.config=classpath:ehcache.xml
